[FLINK-37703][hadoop] Splitting the non-local write unit test to seperated class to fix the unstable tests
